+++ api/src/main/java/jakarta/data/metamodel/range/impl/NullValue.java
@@ -15,41 +15,14 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.metamodel.range;
+package jakarta.data.metamodel.range.impl;
 
+import jakarta.data.metamodel.range.Range;
 import jakarta.data.metamodel.restrict.Operator;
 
-import java.util.Objects;
-
-public record LowerBound<T extends Comparable<T>>(T bound, boolean strict)
-        implements Range<T> {
-    public LowerBound {
-        Objects.requireNonNull(bound, "Lower bound must not be null");
-    }
-
-    public LowerBound(T bound) {
-        this(bound, false);
-    }
-
+public record NullValue<T>() implements Range<T> {
     @Override
     public Operator operator() {
-        return strict ? Operator.GREATER_THAN : Operator.GREATER_THAN_EQUAL;
-    }
-
-    @Override
-    public String toString() {
-        return bound.toString();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        return obj instanceof LowerBound<?> that
-            && bound.equals(that.bound)
-            && strict == that.strict;
-    }
-
-    @Override
-    public int hashCode() {
-        return bound.hashCode();
+        return Operator.NULL;
     }
 }
