+++ api/src/main/java/jakarta/data/metamodel/range/impl/UpperBound.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+package jakarta.data.metamodel.range.impl;
+
+import jakarta.data.metamodel.range.Range;
+import jakarta.data.metamodel.restrict.Operator;
+
+import java.util.Objects;
+
+public record UpperBound<T extends Comparable<T>>(T bound, boolean strict)
+        implements Range<T> {
+    public UpperBound {
+        Objects.requireNonNull(bound, "Upper bound must not be null");
+    }
+
+    public UpperBound(T bound) {
+        this(bound, false);
+    }
+
+    @Override
+    public Operator operator() {
+        return strict ? Operator.LESS_THAN : Operator.LESS_THAN_EQUAL;
+    }
+
+    @Override
+    public String toString() {
+        return bound.toString();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        return obj instanceof UpperBound<?> that
+            && bound.equals(that.bound)
+            && strict == that.strict;
+    }
+
+    @Override
+    public int hashCode() {
+        return bound.hashCode();
+    }
+}
