+++ api/src/main/java/module-info.java
@@ -1116,6 +1116,7 @@ module jakarta.data {
     exports jakarta.data.metamodel.impl;
     exports jakarta.data.metamodel.restrict;
     exports jakarta.data.metamodel.range;
+    exports jakarta.data.metamodel.range.impl;
     exports jakarta.data.page;
     exports jakarta.data.page.impl;
     exports jakarta.data.repository;
